Adds NPC prefab w/ movement and sensor scripts
A new NPC prefab instance is in NPCScene. Currently, the NPC darts to the player when the player enters the sensor, and moves in a jittery manner when following the player. Chase and stop distance values might need adjusting, and the speed of the NPC may need adjusting to match player acceleration  in regards to horizontal and vertical input in the PlayerMovement script.

Player and NPC separation must also be tested, likely with platforms that only the player can jump on so the NPC can't follow the player.