Fixes jittery NPC movement
Normalization is applied to NPC's movement direction for smoother movement. However, since the player moves faster when moving diagonally, the player can out-speed the NPC during diagonal movement, causing the NPC to lose its target. This is because the player's movement direction is not normalized. This can be fixed by normalizing the move vector in the player's Update function, but this also removes the player's smooth acceleration/deceleration. Using a different input method can fix this, or adjusting the NPC's move direction vector in its Update function can also solve the issue.